refactor: Blog component'i screens'e tasindi.
