+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 
 import Main from "./screens/Main";
-import Blog from "./components/blog/Blog";
+import Blog from "./screens/Blog";
 import BlogPost from "./screens/BlogPost";
 import Footer from "./components/Footer";
 import NotFound from "./components/NotFound";
