--- src/components/blog/Blog.css
@@ -1,53 +1,0 @@
-.component-blog {
-  min-height: 85vh;
-  background-image: linear-gradient(#0c0b0b, #060606);
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.component-blog .list-group-item {
-  margin: 0, 5px;
-  background-color: transparent;
-  border: 1px solid white;
-  color: white;
-  border-radius: 100px;
-}
-
-.component-blog .Categories {
-    position: relative;
-    right: 115px;
-      
-}
-
-.logo {
-  max-width: 50%;
-}
-
-.logo p {
-  color: white !important; /* temporary solution: LinkContainer makes it blue */
-}
-
-.list-group-item {
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-}
-
-.blog-card h4 {
-    font-size: calc(5px + 2vmin);
-    color: #006395;
-}
-
-.blog-card p {
-    font-size: calc(4px + 2vmin);
-    color: white;
-}
-
-.blog-card span {
-    font-size: calc(1px + 2vmin);
-    color: #1F3C65;
-}
-
-
